unload agent clients

+++ main.py
@@ -21,7 +21,7 @@ from doc_control import check_rashod
 
 def auto_load(prm_cursor):
     white_list = []
-    load_all = 1
+    load_all = 0
     if load_all == 1:
         white_list.append(3716)  # расходнаядоставка
         white_list.append(410)  # расходнаянакладная
@@ -45,10 +45,12 @@ def auto_load(prm_cursor):
 
         white_list.append(5468)  # производители импортеры
         white_list.append(5196)  # даты розлива
+        white_list.append(4840)  # клиенты агента
 
     if load_all == 0:
         pass
-        white_list.append(3716)  #
+        white_list.append(4840)  # клиенты агента
+        # white_list.append(3716)  #
         # white_list.append(434)  #
         # white_list.append(434)  # приход
         # white_list.append(4308)  # выручкадоставка  sp4323 переброска
@@ -138,9 +140,10 @@ def auto_load(prm_cursor):
                 # приходы
                 prihod.load_prihod(prm_cursor, wsdl_client, row_delta)
             try:
-                prm_cursor.execute('''delete from _1SUPDTS where (DBSIGN = 'P1 ') and (DWNLDID='1122!!')
-                                    and (OBJID=%s) and (TYPEID=%s)''', (row_delta['OBJID'], row_delta['TYPEID']))
-                conn.commit()
+                if row_delta['TYPEID'] != 4840:
+                    prm_cursor.execute('''delete from _1SUPDTS where (DBSIGN = 'P1 ') and (DWNLDID='1122!!')
+                                        and (OBJID=%s) and (TYPEID=%s)''', (row_delta['OBJID'], row_delta['TYPEID']))
+                    conn.commit()
                 logging.info(';'.join(['Загружен объект', str(row_delta['OBJID']), str(row_delta['TYPEID'])]))
             except:
                 logging.error(';'.join(['Ошибка загрузки объекта', str(row_delta['OBJID']),
